modify properties
